AI Bank Seat Book/requirements.txt
Flask==2.3.3
deepface==0.0.78
Pillow==9.5.0